+++ src/thinger/system/platform/backup.h
@@ -8,7 +8,6 @@
 #include "../../utils/docker.h"
 #include "../../utils/tar.h"
 
-
 class PlatformBackup : public ThingerMonitorBackup {
 
 public:
@@ -24,30 +23,68 @@ public:
 
         file_to_upload = name_+"_"+tag_+".tar.gz";
 
-        create_backup_folder();
     }
 
-    void create() {
+    json backup() {
 
-        backup_thinger();
-        backup_mongodb();
-        backup_influxdb();
-        backup_plugins();
-        compress_backup();
+        json data;
+        data["operation"] = {};
+        data["operation"]["create_backup_folder"] = create_backup_folder();
+        if (!data["operation"]["create_backup_folder"]["status"].get<bool>()) {
+            data["status"] = false;
+            return data;
+        }
 
+        data["operation"]["backup_thinger"]   = backup_thinger();
+        data["operation"]["backup_mongodb"]   = backup_mongodb();
+        data["operation"]["backup_influxdb"]  = backup_influxdb();
+        data["operation"]["backup_plugins"]   = backup_plugins();
+        data["operation"]["compress_backup"]  = compress_backup();
+
+        // Set global status value
+        data["status"] = true;
+        for (auto& element : data["operation"]) {
+            if (!element["status"].get<bool>()) {
+                data["status"] = false;
+                break;
+            }
+        }
+
+        return data;
     }
 
-    int upload()  {
+    json upload()  {
+
+        json data;
 
         if ( storage == "S3" )
-            return AWS::multipart_upload_to_s3(backup_folder+"/"+file_to_upload, bucket, region, access_key, secret_key);
-            //return AWS::upload_to_s3(backup_folder+"/"+file_to_upload, bucket, region, access_key, secret_key);
+            data["operation"]["upload_s3"] = {};
+            if(AWS::multipart_upload_to_s3(backup_folder+"/"+file_to_upload, bucket, region, access_key, secret_key))
+                data["operation"]["upload_s3"]["status"] = true;
+            else {
+                data["operation"]["upload_s3"]["status"] = false;
+                data["operation"]["upload_s3"]["error"].push_back("Failed uploading to S3");
+            }
         // else if
-        return -1;
+
+        data["status"] = true;
+        for (auto& element : data["operation"]) {
+            if (!element["status"].get<bool>()) {
+                data["status"] = false;
+                break;
+            }
+        }
+
+        return data;
     }
 
-    void clean() {
-        clean_thinger();
+    json clean() {
+
+        json data;
+
+        clean_backup() ? data["status"] = true : data["status"] = false;
+
+        return data;
     }
 
 protected:
@@ -62,22 +99,33 @@ protected:
     std::string file_to_upload;
 
 private:
-    void create_backup_folder() {
+
+   json create_backup_folder() {
+        json data;
+
         std::filesystem::remove_all(backup_folder+"/"+tag_);
-        std::filesystem::create_directories(backup_folder+"/"+tag_);
+        if (!std::filesystem::create_directories(backup_folder+"/"+tag_)) {
+          data["status"]  = false;
+          data["error"].push_back("Failed to create backup directory");
+        } else
+          data["status"]  = true;
+
+        return data;
     }
 
-    void compress_backup() {
-        Tar::create(backup_folder+"/"+tag_, backup_folder+"/"+file_to_upload);
-    }
+    json backup_thinger() {
+        json data;
 
-    void backup_thinger() {
         // With tar creation instead of copying to folder we maintain ownership and permissions
+        data["status"] = true;
         if (std::filesystem::exists(config_.get_backups_data_path()+"/thinger/users"))
-            Tar::create(config_.get_backups_data_path()+"/thinger/users", backup_folder+"/"+tag_+"/thinger-"+tag_+".tar");
+            data["status"] = Tar::create(config_.get_backups_data_path()+"/thinger/users", backup_folder+"/"+tag_+"/thinger-"+tag_+".tar");
+
+        return data;
     }
 
-    void backup_mongodb() {
+    json backup_mongodb() {
+        json data;
         // get mongodb root password
         std::ifstream compose (config_.get_backups_compose_path()+"/docker-compose.yml", std::ifstream::in);
         std::string line;
@@ -92,11 +140,26 @@ private:
             }
         }
 
-        Docker::Container::exec("mongodb", "mongodump -u thinger -p "+mongo_password);
-        Docker::Container::copy_from_container("mongodb", "/dump", backup_folder+"/"+tag_+"/mongodbdump-"+tag_+".tar");
+        if (!Docker::Container::exec("mongodb", "mongodump -u thinger -p "+mongo_password)) {
+            data["status"]  = false;
+            data["error"].push_back("Failed executing mongodb backup");
+            return data;
+        }
+
+        if (!Docker::Container::copy_from_container("mongodb", "/dump", backup_folder+"/"+tag_+"/mongodbdump-"+tag_+".tar")) {
+            data["status"]  = false;
+            data["error"].push_back("Failed copying mongodb backup from container");
+            return data;
+        }
+
+        data["status"] = true;
+
+        return data;
     }
 
-    void backup_influxdb() {
+    json backup_influxdb() {
+        json data;
+
         if (std::filesystem::exists(config_.get_backups_data_path()+"/influxdb2")) {
             // get influx token
             std::ifstream compose (config_.get_backups_compose_path()+"/docker-compose.yml", std::ifstream::in);
@@ -111,43 +174,102 @@ private:
                     influx_token = line.substr(first_del+1, last_del - first_del-1);
                 }
             }
-            Docker::Container::exec("influxdb2", "influx backup /dump -t "+influx_token);
-            Docker::Container::copy_from_container("influxdb2", "/dump", backup_folder+"/"+tag_+"/influxdb2dump-"+tag_+".tar");
+            if (!Docker::Container::exec("influxdb2", "influx backup /dump -t "+influx_token)) {
+                data["status"]  = false;
+                data["error"].push_back("Failed executing influxdb2 backup");
+                return data;
+            }
+            if (!Docker::Container::copy_from_container("influxdb2", "/dump", backup_folder+"/"+tag_+"/influxdb2dump-"+tag_+".tar")) {
+                data["status"]  = false;
+                data["error"].push_back("Failed copying influxdb2 backup from container");
+                return data;
+            }
         } else {
-            Docker::Container::exec("influxdb", "influxd backup --portable /dump");
-            Docker::Container::copy_from_container("influxdb", "/dump", backup_folder+"/"+tag_+"/influxdbdump-"+tag_+".tar");
+            if (!Docker::Container::exec("influxdb", "influxd backup --portable /dump")) {
+                data["status"]  = false;
+                data["error"].push_back("Failed executing influxdb backup");
+                return data;
+            }
+            if (!Docker::Container::copy_from_container("influxdb", "/dump", backup_folder+"/"+tag_+"/influxdbdump-"+tag_+".tar")) {
+                data["status"]  = false;
+                data["error"].push_back("Failed copying influxdb backup from container");
+                return data;
+            }
         }
+
+        data["status"] = true;
+
+        return data;
+
     }
 
-    void backup_plugins() {
+    json backup_plugins() {
 
-        std::filesystem::create_directories(backup_folder+"/"+tag_+"/plugins");
+        json data;
 
-        if (! std::filesystem::exists(config_.get_backups_data_path()+"/thinger/users")) return;
+        if (!std::filesystem::create_directories(backup_folder+"/"+tag_+"/plugins")) {
+            data["status"]  = false;
+            data["error"].push_back("Failed creating plugins directory in backup folder");
+            return data;
+        }
+
+        if (! std::filesystem::exists(config_.get_backups_data_path()+"/thinger/users")) {
+            data["status"] = true;
+            data["msg"].push_back("Platform has no users");
+            return data;
+        }
 
         for (const auto & p1 : fs::directory_iterator(config_.get_backups_data_path()+"/thinger/users/")) { // users
             if (! std::filesystem::exists(p1.path().string()+"/plugins/")) continue;
 
+            std::string user = p1.path().filename().string();
+
             // backups networks
-            Docker::Network::inspect(p1.path().filename().string(), backup_folder+"/"+tag_+"/plugins");
+            if ( Docker::Network::inspect(user, backup_folder+"/"+tag_+"/plugins") ) {
+                data["msg"].push_back("Backed up "+user+" docker network");
+            } else {
+                data["error"].push_back("Failed backing up "+user+" docker network");
+            }
 
             // backup plugins
             for (const auto & p2 : fs::directory_iterator(p1.path().string()+"/plugins/")) { // plugins
-                std::string container_name = p1.path().filename().string()+"-"+p2.path().filename().string();
-                Docker::Container::inspect(container_name, backup_folder+"/"+tag_+"/plugins");
+                std::string container_name = user+"-"+p2.path().filename().string();
+                if (Docker::Container::inspect(container_name, backup_folder+"/"+tag_+"/plugins")) {
+                    data["msg"].push_back("Backed up "+container_name+" docker container");
+                } else {
+                    data["error"].push_back("Failed backing up "+container_name+" docker container");
+                }
             }
         }
+
+        if (data.contains("error"))
+            data["status"] = false;
+        else
+            data["status"] = true;
+
+        return data;
     }
 
-    void clean_thinger() {
-        std::filesystem::remove_all(backup_folder+"/"+tag_);
-        std::filesystem::remove_all(backup_folder+"/"+file_to_upload);
-        Docker::Container::exec("mongodb", "rm -rf /dump");
+    json compress_backup() {
+        json data;
+
+        data["status"] = Tar::create(backup_folder+"/"+tag_, backup_folder+"/"+file_to_upload);
+
+        return data;
+    }
+
+    bool clean_backup() {
+        bool status = true;
+        status = status && std::filesystem::remove_all(backup_folder+"/"+tag_);
+        status = status && std::filesystem::remove_all(backup_folder+"/"+file_to_upload);
+        status = status && Docker::Container::exec("mongodb", "rm -rf /dump");
         if (std::filesystem::exists(config_.get_backups_data_path()+"/influxdb2")) {
-            Docker::Container::exec("influxdb2", "rm -rf /dump");
+            status = status && Docker::Container::exec("influxdb2", "rm -rf /dump");
         } else {
-            Docker::Container::exec("influxdb", "rm -rf /dump");
+            status = status && Docker::Container::exec("influxdb", "rm -rf /dump");
         }
+
+        return status;
     }
 
 };